aws-dynamodb-poc/开发/crud_dynamodb/requirements.txt
boto3==1.26.165
python-dotenv==0.21.1
